mediatek/packages/apps/RCSe/core/src/com/mediatek/rcse/service/binder/ChatEventStructForBinder.java
/* Copyright Statement:
 *
 * This software/firmware and related documentation ("MediaTek Software") are
 * protected under relevant copyright laws. The information contained herein
 * is confidential and proprietary to MediaTek Inc. and/or its licensors.
 * Without the prior written permission of MediaTek inc. and/or its licensors,
 * any reproduction, modification, use or disclosure of MediaTek Software,
 * and information contained herein, in whole or in part, shall be strictly prohibited.
 *
 * MediaTek Inc. (C) 2012. All rights reserved.
 *
 * BY OPENING THIS FILE, RECEIVER HEREBY UNEQUIVOCALLY ACKNOWLEDGES AND AGREES
 * THAT THE SOFTWARE/FIRMWARE AND ITS DOCUMENTATIONS ("MEDIATEK SOFTWARE")
 * RECEIVED FROM MEDIATEK AND/OR ITS REPRESENTATIVES ARE PROVIDED TO RECEIVER ON
 * AN "AS-IS" BASIS ONLY. MEDIATEK EXPRESSLY DISCLAIMS ANY AND ALL WARRANTIES,
 * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED WARRANTIES OF
 * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE OR NONINFRINGEMENT.
 * NEITHER DOES MEDIATEK PROVIDE ANY WARRANTY WHATSOEVER WITH RESPECT TO THE
 * SOFTWARE OF ANY THIRD PARTY WHICH MAY BE USED BY, INCORPORATED IN, OR
 * SUPPLIED WITH THE MEDIATEK SOFTWARE, AND RECEIVER AGREES TO LOOK ONLY TO SUCH
 * THIRD PARTY FOR ANY WARRANTY CLAIM RELATING THERETO. RECEIVER EXPRESSLY ACKNOWLEDGES
 * THAT IT IS RECEIVER'S SOLE RESPONSIBILITY TO OBTAIN FROM ANY THIRD PARTY ALL PROPER LICENSES
 * CONTAINED IN MEDIATEK SOFTWARE. MEDIATEK SHALL ALSO NOT BE RESPONSIBLE FOR ANY MEDIATEK
 * SOFTWARE RELEASES MADE TO RECEIVER'S SPECIFICATION OR TO CONFORM TO A PARTICULAR
 * STANDARD OR OPEN FORUM. RECEIVER'S SOLE AND EXCLUSIVE REMEDY AND MEDIATEK'S ENTIRE AND
 * CUMULATIVE LIABILITY WITH RESPECT TO THE MEDIATEK SOFTWARE RELEASED HEREUNDER WILL BE,
 * AT MEDIATEK'S OPTION, TO REVISE OR REPLACE THE MEDIATEK SOFTWARE AT ISSUE,
 * OR REFUND ANY SOFTWARE LICENSE FEES OR SERVICE CHARGE PAID BY RECEIVER TO
 * MEDIATEK FOR SUCH MEDIATEK SOFTWARE AT ISSUE.
 *
 * The following software/firmware and/or related documentation ("MediaTek Software")
 * have been modified by MediaTek Inc. All revisions are subject to any receiver's
 * applicable license agreements with MediaTek Inc.
 */

package com.mediatek.rcse.service.binder;

import android.os.Parcel;
import android.os.Parcelable;

import java.util.ArrayList;
import java.util.List;

/**
 * This class represents a chat event structure used to be shared
 */
public class ChatEventStructForBinder implements Parcelable {
    public static final String TAG = "ChatEventStruct";

    public ChatEventStructForBinder(int info, List<String> relatedInfo, String dt) {
        mInformation = info;
        mRelatedInformation = relatedInfo;
        mDate = dt;
    }

    public ChatEventStructForBinder(Parcel source) {
        mRelatedInformation = new ArrayList<String>();
        mInformation = source.readInt();
        source.readStringList(mRelatedInformation);
        mDate = source.readString();
    }

    private int mInformation = 0;

    private List<String> mRelatedInformation;

    private String mDate = null;

    /**
     * To string
     * 
     * @return information
     */
    public String toString() {
        return TAG + "information is " + mInformation + " relatedInformation is " + mRelatedInformation + " date is "
                + mDate;
    }

    @Override
    public int describeContents() {
        return 0;
    }

    @Override
    public void writeToParcel(Parcel dest, int flags) {
        mRelatedInformation = new ArrayList<String>();
        dest.writeInt(mInformation);
        dest.writeStringList(mRelatedInformation);
        dest.writeString(mDate);
    }

    /**
     * Parcelable creator
     */
    public static final Parcelable.Creator<ChatEventStructForBinder> CREATOR =
            new Parcelable.Creator<ChatEventStructForBinder>() {

        @Override
        public ChatEventStructForBinder createFromParcel(Parcel source) {
                    return new ChatEventStructForBinder(source);
        }

        public ChatEventStructForBinder[] newArray(int size) {
            return new ChatEventStructForBinder[size];
        }
    };
}
